Basic implementation of the database user oauth2 login from web project
This creates a basic implementation of UserDetails in order to take the
login credentials from the cores of the REST API, next step is to
connect this logic with the current module we have with mongo.

In web security config I changed some code in order to work with an
object instead of burned memory users.